Update quick-score.js with code from quickey-quick-score.js
Clean up order of params in adjustRemainingScore().
Include exports from config in index.js.
Update scoreArray() to not default the config to null, so that DefaultConfig gets used.
Remove quickey-quick-score.js.
Fix index.test.js with new arity for quickScore().
Update .eslintrc.
