+++ src/index.js
@@ -1,4 +1,4 @@
 export {quickScore} from "./quick-score";
-export {quickeyQuickScore} from "./quickey-quick-score";
+export * from "./config";
 export {Range} from "./range";
 export {scoreArray, createScorer} from "./score-array";
