+++ src/quick-score.js
@@ -1,145 +1,131 @@
 import {Range} from "./range";
+import {DefaultConfig} from "./config";
 
 
-const WordSeparators = "-/\\:()<>%.,_=&[] \t\n\r";
-const UpperCaseLetters = (function() {
-	const charCodeA = "A".charCodeAt(0);
-	const uppercase = [];
-
-	for (let i = 0; i < 26; i++) {
-		uppercase.push(String.fromCharCode(charCodeA + i));
-	}
-
-	return uppercase.join("");
-})();
-const IgnoredScore = 0.9;
-const SkippedScore = 0.15;
 const LongStringLength = 151;
-//const MaxMatchStartPct = .15;
-const MinMatchDensityPct = .75;
-const MaxMatchDensityPct = .95;
-const BeginningOfStringPct = .1;
+const MaxMatchStartPct = .15;
 
 
 export function quickScore(
 	string,
 	query,
 	matches,
-	noSkipReduction,
-	searchRange = new Range(0, string.length),
-	queryRange = new Range(0, query.length),
-	fullMatchedRange = new Range())
+	config = DefaultConfig,
+	stringRange = new Range(0, string.length))
 {
-	if (!queryRange.length) {
-			// deduct some points for all remaining characters
-		return IgnoredScore;
+	if (!query) {
+		return config.skippedScore;
 	}
 
-	if (queryRange.length > searchRange.length) {
-		return 0;
-	}
+	const lcString = string.toLocaleLowerCase();
+	const lcQuery = query.toLocaleLowerCase();
 
-	for (let i = queryRange.length; i > 0; i--) {
-		const querySubstring = query.substring(queryRange.location, queryRange.location + i);
-			// reduce the length of the search range by the number of chars
-			// we're skipping in the query, to make sure there's enough string
-			// left to possibly contain the skipped chars
-		const matchedRange = getRangeOfSubstring(string, querySubstring,
-			new Range(searchRange.location, searchRange.length - queryRange.length + i));
+	return calcScore(stringRange, new Range(0, query.length), new Range());
 
-		if (!matchedRange.isValid()) {
-			continue;
+
+	function calcScore(
+		searchRange,
+		queryRange,
+		fullMatchedRange)
+	{
+		if (!queryRange.length) {
+				// deduct some points for all remaining characters
+			return config.ignoredScore;
 		}
 
-		if (!fullMatchedRange.isValid()) {
-			fullMatchedRange.location = matchedRange.location;
-		} else {
-			fullMatchedRange.location = Math.min(fullMatchedRange.location, matchedRange.location);
+		if (queryRange.length > searchRange.length) {
+			return 0;
 		}
 
-		fullMatchedRange.max(matchedRange.max());
+		for (let i = queryRange.length; i > 0; i--) {
+			const querySubstring = lcQuery.substring(queryRange.location, queryRange.location + i);
+				// reduce the length of the search range by the number of chars
+				// we're skipping in the query, to make sure there's enough string
+				// left to possibly contain the skipped chars
+			const matchedRange = getRangeOfSubstring(lcString, querySubstring,
+				new Range(searchRange.location, searchRange.length - queryRange.length + i));
 
-		if (matches) {
-			matches.push([matchedRange.location, matchedRange.max()]);
-		}
-
-		const remainingSearchRange = new Range(matchedRange.max(), searchRange.max() - matchedRange.max());
-		const remainingQueryRange = new Range(queryRange.location + i, queryRange.length - i);
-		const remainingScore = quickScore(string, query, matches,
-			noSkipReduction, remainingSearchRange, remainingQueryRange, fullMatchedRange);
-
-		if (remainingScore) {
-			const matchStartPercentage = fullMatchedRange.location / string.length;
-			const isShortString = string.length < LongStringLength;
-			const useSkipReduction = true;
-//			const useSkipReduction = !noSkipReduction && (isShortString || matchStartPercentage < MaxMatchStartPct),
-			let score = remainingSearchRange.location - searchRange.location;
-			let matchStartDiscount = (1 - matchStartPercentage);
-				// default to no match-sparseness discount, for cases where there
-				// are spaces before the matched letters or they're capitals
-			let matchRangeDiscount = 1;
-
-			if (matchedRange.location > searchRange.location) {
-					// some letters were skipped when finding this match, so
-					// adjust the score based on whether spaces or capital
-					// letters were skipped
-				if (useSkipReduction && WordSeparators.indexOf(string.charAt(matchedRange.location - 1)) > -1) {
-					for (let j = matchedRange.location - 2; j >= searchRange.location; j--) {
-						if (WordSeparators.indexOf(string.charAt(j)) > -1) {
-							score--;
-						} else {
-							score -= SkippedScore;
-						}
-					}
-				} else if (useSkipReduction && UpperCaseLetters.indexOf(string.charAt(matchedRange.location)) > -1) {
-					for (let j = matchedRange.location - 1; j >= searchRange.location; j--) {
-						if (UpperCaseLetters.indexOf(string.charAt(j)) > -1) {
-							score--;
-						} else {
-							score -= SkippedScore;
-						}
-					}
-				} else {
-						// reduce the score by the number of chars we've
-						// skipped since the beginning of the search range
-						// and discount the remainingScore based on how much
-						// larger the match is than the query, unless
-						// the match is in the first 10% of the string, the
-						// match range isn't too sparse and the whole string
-						// is not too long
-//					score -= matchedRange.location - searchRange.location;
-					score -= (matchedRange.location - searchRange.location) / 2.0;
-
-					matchRangeDiscount = query.length / fullMatchedRange.length;
-					matchRangeDiscount = (isShortString &&
-						matchStartPercentage <= BeginningOfStringPct &&
-						matchRangeDiscount >= MinMatchDensityPct) ? 1 : matchRangeDiscount;
-					matchStartDiscount = matchRangeDiscount >= MaxMatchDensityPct ?
-						1 : matchStartDiscount;
-				}
+			if (!matchedRange.isValid()) {
+				continue;
 			}
 
-				// discount the scores of very long strings
-//			score += remainingScore * Math.min(remainingSearchRange.length, LongStringLength) *
-//			score += remainingScore * remainingSearchRange.length *
-//				matchRangeDiscount * matchStartDiscount;
-			score += remainingScore * remainingSearchRange.length;
+			if (!fullMatchedRange.isValid()) {
+				fullMatchedRange.location = matchedRange.location;
+			} else {
+				fullMatchedRange.location = Math.min(fullMatchedRange.location, matchedRange.location);
+			}
 
-			score /= searchRange.length;
+			fullMatchedRange.max(matchedRange.max());
 
-			return score;
+			if (matches) {
+				matches.push([matchedRange.location, matchedRange.max()]);
+			}
+
+			const remainingSearchRange = new Range(matchedRange.max(), searchRange.max() - matchedRange.max());
+			const remainingQueryRange = new Range(queryRange.location + i, queryRange.length - i);
+			const remainingScore = calcScore(remainingSearchRange, remainingQueryRange, fullMatchedRange);
+
+			if (remainingScore) {
+				const isShortString = string.length < LongStringLength;
+				const matchStartPercentage = fullMatchedRange.location / string.length;
+				const useSkipReduction = config.skipReduction === true &&
+					(isShortString || matchStartPercentage < MaxMatchStartPct);
+				let score = remainingSearchRange.location - searchRange.location;
+					// default to true since we only want to apply a discount if
+					// we hit the final else clause below, and we won't get to
+					// any of them if the match is right at the start of the
+					// searchRange
+				let skippedSpecialChar = true;
+
+				if (matchedRange.location > searchRange.location) {
+						// some letters were skipped when finding this match, so
+						// adjust the score based on whether spaces or capital
+						// letters were skipped
+					if (useSkipReduction &&
+							config.wordSeparators.indexOf(string[matchedRange.location - 1]) > -1) {
+						for (let j = matchedRange.location - 2; j >= searchRange.location; j--) {
+							if (config.wordSeparators.indexOf(string[j]) > -1) {
+								score--;
+							} else {
+								score -= config.skippedScore;
+							}
+						}
+					} else if (useSkipReduction &&
+							config.uppercaseLetters.indexOf(string[matchedRange.location]) > -1) {
+						for (let j = matchedRange.location - 1; j >= searchRange.location; j--) {
+							if (config.uppercaseLetters.indexOf(string[j]) > -1) {
+								score--;
+							} else {
+								score -= config.skippedScore;
+							}
+						}
+					} else {
+							// reduce the score by the number of chars we've
+							// skipped since the beginning of the search range
+						score -= matchedRange.location - searchRange.location;
+						skippedSpecialChar = false;
+					}
+				}
+
+				score += config.adjustRemainingScore(string,
+					query, remainingScore, skippedSpecialChar, searchRange,
+					remainingSearchRange, matchedRange, fullMatchedRange);
+				score /= searchRange.length;
+
+				return score;
+			}
 		}
-	}
 
-	if (matches) {
-			// the remaining query does not appear in the remaining
-			// string, so clear the matches, since we'll start over with a
-			// shorter piece of the query, which might match earlier
-			// in the string, making any existing match indexes invalid.
-		matches.length = 0;
-	}
+		if (matches) {
+				// the remaining query does not appear in the remaining
+				// string, so clear the matches, since we'll start over with a
+				// shorter piece of the query, which might match earlier
+				// in the string, making any existing match indexes invalid.
+			matches.length = 0;
+		}
 
-	return 0;
+		return 0;
+	}
 }
 
 
@@ -148,8 +134,8 @@ function getRangeOfSubstring(
 	substring,
 	searchRange = new Range(0, string.length))
 {
-	const stringToSearch = string.substring(searchRange.location, searchRange.max()).toLocaleLowerCase();
-	const subStringIndex = stringToSearch.indexOf(substring.toLocaleLowerCase());
+	const stringToSearch = string.substring(searchRange.location, searchRange.max());
+	const subStringIndex = stringToSearch.indexOf(substring);
 	const result = new Range();
 
 	if (subStringIndex > -1) {
